[CHORE] #35 : move photoPicker to common/utils
